requirements.txt
asgiref==3.8.1
coverage==7.6.10
Django==5.1.5
django-model-utils==5.0.0
django-stubs==5.1.2
django-stubs-ext==5.1.2
djangorestframework==3.15.2
flake8==7.1.1
isort==6.0.0
mccabe==0.7.0
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pyflakes==3.2.0
python-decouple==3.8
sqlparse==0.5.3
types-PyYAML==6.0.12.20241230
typing_extensions==4.12.2
